pyproject.toml
[project]
name = "whisper-transcriber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "certifi>=2025.7.14",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "mongo>=0.2.0",
    "openai-whisper>=20250625",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
